refactor: widen installed_channels bitmask to u64
The channel set reached 31 of 32 bits (one free), so the next channel would
overflow. Widen the CH_* bits, ChannelDef.bit, and Cov.installed_channels from
u32 to u64 (63 usable bits) before adding thread-pool timers. Dropping Cov._pad
keeps the section size identical (old u32 + u32 pad == new u64 == 8 bytes) and
the low bits still hold channels 0..N, so the shared mech<->hook layout is
binary-unchanged. Protocol/wire is untouched (installed never goes on the wire;
mech translates bits to channel names).

Pure capacity refactor, no new channel: 38 tests, clippy clean, both targets,
harness 28 PASS x64+x86 (regression - no channel changed).

+++ crates/mech/src/lib.rs
@@ -281,7 +281,7 @@ unsafe fn core_is_alive(pid: u32) -> bool {
 ///
 /// # Safety
 /// `cov` must point to a live, correctly aligned `Cov`.
-unsafe fn gather_coverage(cov: *const Cov, installed: u32, scale_duration: bool) -> Coverage {
+unsafe fn gather_coverage(cov: *const Cov, installed: u64, scale_duration: bool) -> Coverage {
     let mut out = Coverage::default();
     // Track which KIND of observed channel actually ran, so the audit names the right reason: an
     // object wait left real (class B) vs a multimedia timer left real (class C, winmm/ADR-2).
